+++ app/src/main/assets/src/html/onsPages/addPassword.html
@@ -18,8 +18,14 @@
         <ons-list-item modifier="nodivider">
             <ons-input id="username" modifier="underbar" placeholder="Loginname" float></ons-input>
         </ons-list-item>
-        <ons-list-item>
-            <ons-input id="password" modifier="underbar" placeholder="Passwort" type="password" value="" float></ons-input>
+        <ons-list-item modifier="nodivider">
+            <ons-input id="password" modifier="underbar" placeholder="Passwort" type="password" float></ons-input>
+            <p>
+                <ons-checkbox input-id="passwordCheckbox"></ons-checkbox>
+                <label for="passwordCheckbox">
+                    Passwort anzeigen
+                </label>
+            </p>
         </ons-list-item>
         <ons-list-item style="padding-right: 8px">
             <ons-button id="commitButton" modifier="large">Bestätigen</ons-button>
