+++ app/src/main/assets/src/html/onsPages/listview.html
@@ -12,7 +12,7 @@
     <div id="overview"></div>
   </ons-list>
   <ons-fab position="bottom right" id="addButton">
-    <ons-icon icon="md-plus" onclick="document.querySelector('#myNavigator').pushPage('onsPages/addPassword.html')"></ons-icon>
+    <ons-icon icon="md-plus"></ons-icon>
   </ons-fab>
 
   <ons-popover cancelable direction="down" id="popover">
@@ -20,7 +20,7 @@
       <ons-list-item id="deleteUser" modifier="tappable">
         Benutzer Löschen
       </ons-list-item>
-      <ons-list-item modifier="tappable" onclick="history.back()">
+      <ons-list-item id="logout" modifier="tappable">
         Abmelden
       </ons-list-item>
     </ons-list>
