+++ app/src/main/assets/src/html/onsPages/passwordView.html
@@ -1,4 +1,4 @@
-<ons-page id="passwordDetailed">
+<ons-page id="passwordView">
     <ons-toolbar>
         <div class="left">
             <ons-back-button>Abbrechen</ons-back-button>
@@ -18,10 +18,10 @@
             Url hinzufügen
         </ons-list-item>
         <ons-list-item modifier="nodivider">
-            <ons-input id="username" modifier="underbar" placeholder="Loginname" float readonly></ons-input>
+            <ons-input id="username" modifier="underbar" placeholder="Loginname" float></ons-input>
         </ons-list-item>
         <ons-list-item modifier="nodivider">
-            <ons-input id="password" modifier="underbar" placeholder="Passwort" type="password" float readonly></ons-input>
+            <ons-input id="password" modifier="underbar" placeholder="Passwort" type="password" float></ons-input>
             <p>
                 <ons-checkbox input-id="passwordCheckbox"></ons-checkbox>
                 <label for="passwordCheckbox">
