+++ app/src/main/assets/src/js/app/sessionHandler.js
@@ -1,0 +1,14 @@
+export function getSessionUser() {
+    return sessionStorage.getItem("user")
+}
+
+export function getSessionPassword() {
+    return sessionStorage.getItem("password")
+}
+
+export function logout(){
+    sessionStorage.clear()
+    history.back()
+}
+
+//TODO: resolve "passwords" use here
